requirements.txt
requests
python-dotenv
matplotlib
seaborn 
pytest
pytest-mock
Flask>=2.0.0
Pillow>=9.0.0
docx2txt>=0.8
PyMuPDF>=1.24.0
Flask==2.3.3
PyPDF2==3.0.1